drfecommerce/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.11.0
click==8.1.7
coverage==7.3.3
Django==5.0
django-js-asset==2.2.0
django-mptt==0.15.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
factory-boy==3.3.0
Faker==21.0.0
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
